.pbm and .pgm now work afaik

+++ src/basic.rs
@@ -36,10 +36,10 @@ pub fn packrow(row: &[u8]) -> Vec<u8> {
     packedrow
 }
 
-pub fn numfromascii(bytes: &[u8]) -> usize {
+pub fn numfromascii(bytes: &[u8]) -> u32 {
     let mut n = 0;
     for b in bytes {
-        n += (*b as usize) - 48;
+        n += (*b as u32) - 48;
         n *= 10;
     }
     n / 10
